+++ src/lspClient/sorbetClient.ts
@@ -263,7 +263,7 @@ export class SorbetClient implements vscode.Disposable {
     const logPrefix =
       this.context.clientManager.clientCount > 1 ? `[${this.workspaceFolder.name}] ` : '';
     if (this.languageClient) {
-      this.context.log.info(`${logPrefix}Ignored start request, already running.`);
+      this.context.log.debug(`${logPrefix}Ignored start request, already running.`);
       return;
     } else if (this.configuration.lspConfigurationType === LspConfigurationType.Disabled) {
       this.context.log.info(`${logPrefix}Ignored start request, disabled by configuration.`);